Squashed commit of the following:

commit 6a0563a5523807e0f4cf3d7717aa0fc548f5fb14
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Jun 30 18:27:57 2020 -0400

    build: Fix unit tests to build and work properly on a 32-bit system

+++ src/network_inspectors/appid/test/appid_http_session_test.cc
@@ -152,8 +152,8 @@ void Profiler::show_stats() { }
 
 MemoryContext::MemoryContext(MemoryTracker&) { }
 MemoryContext::~MemoryContext() { }
-void memory::MemoryCap::update_allocations(unsigned long) { }
-void memory::MemoryCap::update_deallocations(unsigned long) { }
+void memory::MemoryCap::update_allocations(size_t) { }
+void memory::MemoryCap::update_deallocations(size_t) { }
 
 OdpContext::OdpContext(const AppIdConfig&, snort::SnortConfig*) { }
 AppIdConfig::~AppIdConfig() { }
